[BE] refactor: S3Config 사용 Provider 수동 빈 등록으로 변경 #335
