+++ backend/src/main/java/com/onair/hearit/admin/infrastructure/s3/S3Config.java
@@ -1,4 +1,4 @@
-package com.onair.hearit.admin.config;
+package com.onair.hearit.admin.infrastructure.s3;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
@@ -12,10 +12,18 @@ public class S3Config {
     @Value("${aws.region}")
     private String region;
 
+    @Value("${aws.s3.bucket}")
+    private String bucket;
+
     @Bean
     public S3Client s3Client() {
         return S3Client.builder()
                 .region(Region.of(region))
                 .build();
+    }
+
+    @Bean
+    public S3FileProvider s3FileProvider(S3Client s3Client) {
+        return new S3FileProvider(s3Client, bucket);
     }
 }
